account data styling, profile styling, sign out delete moved to settings
