+++ frontend/src/app/account_data/account_data.component.html
@@ -1,20 +1,32 @@
 <div class="middle_box">
     <h1 class="page-heading">Account Data</h1>
-    <mat-form-field class="input-80-width">
+    <div class="column-box">
+    <mat-form-field class="input-box">
         <h2>Name:</h2>
         <input matInput [(ngModel)]="user.name" />
         <button mat-button (click)="saveUsername()">Save</button>
     </mat-form-field>
 
-    <mat-form-field class="input-80-width">
-        <h2>Password:</h2>
+    <mat-form-field class="input-box">
+        <h2>New Password:</h2>
         <input matInput [(ngModel)]="user.password" />
+        <h2>Old Password:</h2>
+        <input matInput [(ngModel)]="user.old_password" />
         <button mat-button (click)="savePassword()">Save</button>
     </mat-form-field>
 
-    <mat-form-field class="input-80-width">
+    <mat-form-field class="input-box">
         <h2>Description:</h2>
         <input matInput [(ngModel)]="user.description" />
         <button mat-button (click)="saveDescription()">Save</button>
     </mat-form-field>
+
+    <div class="flex-row">
+        <!-- Signout button -->
+        <button mat-raised-button class="sign-out" (click)="signOut()">Sign Out</button>
+
+        <!-- Delete button -->
+        <button mat-raised-button class="delete-account" (click)="deleteAction()">Delete</button>
+    </div>
+</div>
 </div>
